big update

+++ src/Traits/MetaActionsTrait.php
@@ -1,16 +1,12 @@
 <?php
 
-namespace Aweram\Metable\Livewire;
+namespace Aweram\Metable\Traits;
 
 use Aweram\Metable\Interfaces\MetaModelInterface;
-use Aweram\Metable\Interfaces\ShouldMetaInterface;
 use Aweram\Metable\Models\Meta;
-use Livewire\Component;
 
-class MetaIndexWire extends Component
+trait MetaActionsTrait
 {
-    public ShouldMetaInterface $model;
-
     public bool $displayData = false;
     public int|null $metaId = null;
 
@@ -39,39 +35,6 @@ class MetaIndexWire extends Component
         ];
     }
 
-    public function render()
-    {
-        $metaClass = config("metable.customMetaModel") ?? Meta::class;
-        $metas = $metaClass::query()
-            ->select("id", "name", "content", "property", "separated")
-            ->where("metable_id", $this->model->id)
-            ->where("metable_type", $this->model::class)
-            ->orderBy("name")
-            ->get();
-        return view('ma::livewire.admin.metas', compact("metas"));
-    }
-
-    public function showCreate(): void
-    {
-        $this->resetFields();
-        $this->displayData = true;
-    }
-
-    public function store(): void
-    {
-        // Валидация
-        $this->validate();
-        $this->model->metas()->create([
-            "name" => $this->name,
-            "content" => $this->content,
-            "property" => $this->property,
-            "separated" => $this->separated ? now() : null,
-        ]);
-
-        session()->flash("metas-success", __("Meta tag successfully added"));
-        $this->closeData();
-    }
-
     public function showEdit(int $metaId): void
     {
         $this->resetFields();
@@ -79,6 +42,9 @@ class MetaIndexWire extends Component
         // Найти тег
         $meta = $this->findMeta();
         if (! $meta) return;
+        // Проверить авторизацию
+        $check = $this->checkAuth("update", $meta);
+        if (! $check) return;
 
         $this->name = $meta->name;
         $this->content = $meta->content;
@@ -93,6 +59,9 @@ class MetaIndexWire extends Component
         // Найти тег
         $meta = $this->findMeta();
         if (! $meta) return;
+        // Проверить авторизацию
+        $check = $this->checkAuth("update", $meta);
+        if (! $check) return;
         // Валидация
         $this->validate();
         $meta->update([
@@ -119,6 +88,9 @@ class MetaIndexWire extends Component
         // Найти тег
         $meta = $this->findMeta();
         if (! $meta) return;
+        // Проверить авторизацию
+        $check = $this->checkAuth("delete", $meta);
+        if (! $check) return;
 
         $this->displayDelete = true;
     }
@@ -134,6 +106,9 @@ class MetaIndexWire extends Component
         // Найти тег
         $meta = $this->findMeta();
         if (! $meta) return;
+        // Проверить авторизацию
+        $check = $this->checkAuth("delete", $meta);
+        if (! $check) return;
 
         $meta->delete();
         session()->flash("metas-success", __("Meta tag successfully deleted"));
@@ -141,12 +116,12 @@ class MetaIndexWire extends Component
         $this->closeDelete();
     }
 
-    private function resetFields(): void
+    protected function resetFields(): void
     {
         $this->reset(["name", "content", "property", "metaId"]);
     }
 
-    private function findMeta(): ?MetaModelInterface
+    protected function findMeta(): ?MetaModelInterface
     {
         $metaClass = config("metable.customMetaModel") ?? Meta::class;
         $meta = $metaClass::find($this->metaId);
@@ -156,5 +131,18 @@ class MetaIndexWire extends Component
             return null;
         }
         return $meta;
+    }
+
+    protected function checkAuth(string $action, MetaModelInterface $meta = null): bool
+    {
+        try {
+            $this->authorize($action, $meta ?? (config("metable.customMetaModel") ?? Meta::class));
+            return true;
+        } catch (\Exception $ex) {
+            session()->flash("metas-error", __("Unauthorized action"));
+            $this->closeData();
+            $this->closeDelete();
+            return false;
+        }
     }
 }
